fix: move back NewProvider constructor to main, providers interface to provider package
